adds b-tree version control optimizations for faster transaction processing

+++ src/journal.cpp
@@ -47,6 +47,7 @@ private:
 	size_t bytes_used;
 	nst::u64 sequence;
 	nst::u64 last_synch;
+	nst::u64 last_check;
 	participants_type participants;
 public:
 	static const std::ios_base::openmode o_mode = std::ios::binary|std::ios::app;
@@ -60,6 +61,7 @@ public:
 	,	bytes_used(0)
 	,	sequence(0)
 	,	last_synch(0)
+	,	last_check(0)
 	{
 	}
 	~journal_state(){
@@ -242,51 +244,54 @@ public:
 			writer.flush();
 			journal_ostr.flush();
 			journal_ostr.rdbuf()->pubsync();
-			Poco::File jf(journal_name);
-			if(jf.exists()){
-				nst::u64 jsize = 0;
-				try{
-					jsize = jf.getSize();
-				}catch(...){
-					printf("error getting journal size\n");
-					return;
-				}
-				set_treestore_journal_size( jsize );
-
-				if(force || jsize > (nst::u64)get_treestore_journal_lower_max() ){
-
-					nst::u64 singles = 0;
-					for(participants_type::iterator p = (*this).participants.begin(); p != (*this).participants.end(); ++p){
-						singles += (*p).second->make_singular() ? 1 : 0;
+			if(force || os::millis() - last_check > 3000){
+				last_check = os::millis();
+				Poco::File jf(journal_name);
+				if(jf.exists()){
+					nst::u64 jsize = 0;
+					try{
+						jsize = jf.getSize();
+					}catch(...){
+						printf("error getting journal size\n");
+						return;
 					}
-					if(singles == participants.size()){
-						set_treestore_journal_size( 0 );
-						printf("journal file > n GB compacting\n");
-						//compact();
-						log_journal(journal_name,"commit",0);
+					set_treestore_journal_size( jsize );
+
+					if(force || jsize > (nst::u64)get_treestore_journal_lower_max() ){
+
+						nst::u64 singles = 0;
 						for(participants_type::iterator p = (*this).participants.begin(); p != (*this).participants.end(); ++p){
-							(*p).second->journal_synch_start();
+							singles += (*p).second->make_singular() ? 1 : 0;
 						}
-						for(participants_type::iterator p = (*this).participants.begin(); p != (*this).participants.end(); ++p){
-							/// test if the data files have not been deleted
-							if(is_valid_storage_directory((*p).second->get_name())){
-								(*p).second->journal_commit();
+						if(singles == participants.size()){
+							set_treestore_journal_size( 0 );
+							printf("journal file > n GB compacting\n");
+							//compact();
+							log_journal(journal_name,"commit",0);
+							for(participants_type::iterator p = (*this).participants.begin(); p != (*this).participants.end(); ++p){
+								(*p).second->journal_synch_start();
 							}
+							for(participants_type::iterator p = (*this).participants.begin(); p != (*this).participants.end(); ++p){
+								/// test if the data files have not been deleted
+								if(is_valid_storage_directory((*p).second->get_name())){
+									(*p).second->journal_commit();
+								}
+							}
+							for(participants_type::iterator p = (*this).participants.begin(); p != (*this).participants.end(); ++p){
+								(*p).second->journal_synch_end();
+							}
+							printf("journal file compacting complete\n");
+							journal_ostr.close();
+							jf.remove();
+							sequence = 0;
+							last_synch = sequence;
+							journal_ostr.open(journal_name, std::ios::binary);
+							set_treestore_journal_size( jf.getSize() );
 						}
-						for(participants_type::iterator p = (*this).participants.begin(); p != (*this).participants.end(); ++p){
-							(*p).second->journal_synch_end();
-						}
-						printf("journal file compacting complete\n");
-						journal_ostr.close();
-						jf.remove();
-						sequence = 0;
-						last_synch = sequence;
-						journal_ostr.open(journal_name, std::ios::binary);
-						set_treestore_journal_size( jf.getSize() );
 					}
 				}
 			}
-		}
+		}  /// last_check
 	}
 };
